pyproject.toml
[project]
name = "research-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-api-python-client>=2.193.0",
    "google-genai>=1.68.0",
    "google-play-scraper>=1.2.7",
    "lxml>=6.0.2",
    "playwright>=1.58.0",
    "praw>=7.8.1",
    "python-dotenv>=1.2.2",
    "pytrends>=4.9.2",
    "requests>=2.32.5",
]
